homework Hibernate about Join - done

+++ src/main/java/EmployeeDAO.java
@@ -1,12 +1,11 @@
 import model.Employee;
-
 import java.util.List;
 
 public interface EmployeeDAO {
     void addEmployee(Employee employee);
-    Employee returnEmployeeById(int id);
+    Employee returnEmployeeById(Long id);
     List<Employee> returnAllEmployee();
-    void updateEmployee(Employee employee, int id);
+    void updateEmployee(Employee employee, Long id);
     void deleteEmployee(Employee employee);
 
 }
